feat: add member map caching service and scheduler?
